cloud_function/PeMS/requirements.txt
cloudevents==1.*
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-pubsub==2.*
protobuf==4.*
